bugfixes in the agent-contest-connector and version control in EILoader
git-svn-id: https://svn.code.sf.net/p/apleis/code/trunk@7 d1be27aa-122a-4001-a9ff-4e8566158a90

+++ src/main/java/eis/examples/acconnector2009/EnvironmentInterface.java
@@ -161,7 +161,7 @@ public class EnvironmentInterface extends EIDefaultImpl implements ConnectionLis
 		// open socket
 		try {
 
-			c = new Connection(this, server.getValue(), ((Long)port.getValue()).intValue());
+			c = new Connection(this, server.getValue(), (port.getValue()).intValue());
 
 		} catch (UnknownHostException e) {
